change-chart-form component
