+++ Frontend/aurum/app/dashboard/page.jsx
@@ -4,6 +4,7 @@ import React from "react";
 import { useState, useEffect } from "react";
 import HamburgerMenu from "../components/dashboard/hamburger-menu";
 import Header from "../components/dashboard/header";
+import ChangeChartForm from "../components/dashboard/change-chart-form";
 import { layouts } from "../../scripts/dashboard_scripts/dashboard_scripts"
 
 export default function DashboardPage() {
@@ -20,15 +21,6 @@ export default function DashboardPage() {
   }, [choosenLayout])
 
 
-  const handleChangeChartClick = (event, segmentIndex, possibleCharts) => {
-    let updatedChoosenCharts = [...choosenCharts]
-    let chartName = event.target.value
-    updatedChoosenCharts[segmentIndex] = possibleCharts.find(c => c.name === chartName)
-    setChoosenCharts(updatedChoosenCharts)
-  }
-
-  console.log(choosenCharts)
-
   return (
     <div className="dashboard page">
       <Header setIsHamburgerOpen={setIsHamburgerOpen} isHamburgerOpen={isHamburgerOpen} />
@@ -40,14 +32,8 @@ export default function DashboardPage() {
           <div key={`${possibleCharts}-${segmentIndex}`} className={`${choosenLayout}-${segmentIndex + 1} chart-container`}>
             {choosenCharts[segmentIndex].chart}
             {isEditMode &&
-              <form className="change-chart-form">
-                <label>change chart</label>
-                <select value={choosenCharts[segmentIndex].name} name="change-chart" onChange={(e) => handleChangeChartClick(e, segmentIndex, possibleCharts)}>
-                  {possibleCharts.map(chart => (
-                    <option name={chart.name} key={chart.name} value={chart.name} >{chart.name}</option>
-                  ))}
-                </select>
-              </form>}
+              <ChangeChartForm choosenCharts={choosenCharts} segmentIndex={segmentIndex} possibleCharts={possibleCharts} setChoosenCharts={setChoosenCharts} />
+            }
           </div>
         ))}
       </div>
